--- data/performance_metrics/7-4-1/balanced_700/simulator_5/config.yaml
@@ -1,4 +1,0 @@
-num_red_steps: 1
-seed: 1
-simulation_time: 20
-time_step: 2
