--- data/performance_metrics/7-4-1/balanced_700/simulator_5/drl/config_1/config.yaml
@@ -1,31 +1,0 @@
-bc_flg: false
-data_augmentation_flg: true
-done_reward: -50
-duration_steps: 5
-features:
-  lane:
-    metric:
-      num_vehicles: true
-    shape:
-      length: true
-      type: true
-  road:
-    metric:
-      delay: true
-      queue_length: true
-  vehicle:
-    direction: true
-    position: true
-    speed: true
-learning_flg: true
-method: apex
-network_id: 1
-num_vehicles: 20
-reward_id: 2
-save_interval: 5
-state_id: 2
-stop:
-  episode: 400
-  flg: true
-  interval: 100
-  type: episode
