--- data/performance_metrics/7-4-1/balanced_700/simulator_5/drl/config_1/data/drl/config_1/config.yaml
@@ -1,55 +1,0 @@
-apex:
-  buffer:
-    batch:
-      number: 1
-      size: 128
-    initial_priority: 500
-    num_new_data: 25
-    priority_reset:
-      episode: 10
-      flg: true
-      interval: 10
-      type: interval
-    size: 200000
-  epsilon:
-    schedule_flg: true
-    value: 0
-  gamma: 0.95
-  learning_rate: 0.0001
-  num_epochs: 1
-  random_action_type: 1
-  td_steps: 1
-  update_interval: 2000
-  weight_decay: 1.0e-05
-drl:
-  bc_flg: false
-  data_augmentation_flg: true
-  done_reward: -50
-  duration_steps: 5
-  features:
-    lane:
-      metric:
-        num_vehicles: true
-      shape:
-        length: true
-        type: true
-    road:
-      metric:
-        delay: true
-        queue_length: true
-    vehicle:
-      direction: true
-      position: true
-      speed: true
-  learning_flg: true
-  method: apex
-  network_id: 1
-  num_vehicles: 20
-  reward_id: 2
-  save_interval: 5
-  state_id: 2
-  stop:
-    episode: 400
-    flg: true
-    interval: 100
-    type: episode
