--- data/performance_metrics/7-4-1/balanced_700/simulator_5/drl/config_1/data/drl/config_1/session/random_phase_probs_df.csv
@@ -1,3 +1,0 @@
-1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17
-0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,0.0588235294117647,,,,,,,,,,,,,,,,,
-,,,,,,,,,,,,,,,,,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705,0.058823529411764705
