--- data/performance_metrics/7-4-1/balanced_700/simulator_5/drl/config_1/data/drl/config_1/session/session_info.json
@@ -1,1 +1,0 @@
-{"update_count": 0}
